+++ index.html
@@ -227,6 +227,6 @@
     </section>
   </template>
   <script src="js/function.js"></script>
-  <script src="js/main.js"></script>
+  <script  type="module" src="js/main.js" ></script>
 </body>
 </html>
