+++ js/seeders/seedersAvatar.js
@@ -1,0 +1,3 @@
+const userAvatar = ['1', '2', '3', '4', '5', '6'];
+
+export {userAvatar}
