+++ js/seeders/seedersDescr.js
@@ -1,0 +1,5 @@
+const userDescr = [
+  'Cool', 'False', 'Nice', 'Great', 'Little', 'Dislike'
+];
+
+export {userDescr}
