+++ js/seeders/seedersLike.js
@@ -1,0 +1,6 @@
+const userLike = [];
+for (let i = 15; i <= 250; i++) {
+  userLike.push(i);
+}
+
+export {userLike}
