+++ js/seeders/seedersName.js
@@ -1,0 +1,4 @@
+const userName = [
+  'Jhon', 'Mike', 'Jessy', 'Alex', 'Anna', 'Paul', 'Jack', 'Caty'
+];
+export {userName};
