+++ js/seeders/seedersUserPhoto.js
@@ -1,0 +1,7 @@
+const userPhotoNumber = [];
+for (let i = 0; i <= 25; i++) {
+  userPhotoNumber.push(i);
+}
+
+
+export {userPhotoNumber};
